+++ rest-server/src/routes/questions.js
@@ -5,8 +5,9 @@ const QuestionsService = require('../services/QuestionsService');
 const checkAuth = require('../middleware/check-auth');
 const notFound = require('../middleware/not-found');
 const badRequest = require('../middleware/bad-request');
+const UserRole = require('../models/user-role');
 
-router.post('/', async (request, response) => {
+router.post('/', checkAuth, async (request, response) => {
   try {
     const question = await QuestionsService.add(request.body);
     response.status(201).json(question);
@@ -15,21 +16,33 @@ router.post('/', async (request, response) => {
   } 
 });
 
-router.get('/', async (request, response) => {
-  const questions = await QuestionsService.getAll();
-  questions && questions.length
-    ? response.json(questions)
-    : response.status(204).end();
+router.get('/', checkAuth, async (request, response) => {
+
+  try {
+    if (UserRole.Admin == request.headers.role) {
+      const questions = await QuestionsService.getAll();
+      questions && questions.length ? response.json(questions) : response.status(204).end();
+    } else {
+      badRequest(request, response, 'This request is not allowed to users with role: USER')
+    }   
+  } catch (error) {
+    badRequest(request, response, error)
+  }
+  
 });
 
-router.get('/:questionId', async (request, response) => {
-  const question = await QuestionsService.getById(request.params.questionId);
-  question
-    ? response.json(question)
-    : notFound(request, response);
+router.get('/:questionId', checkAuth, async (request, response) => {
+
+  try {
+    const question = await QuestionsService.getById(request.params.questionId);
+    question ? response.json(question) : notFound(request, response);
+  } catch (error) {
+    badRequest(request, response, error)
+  }
+  
 });
 
-router.patch('/:questionId', async (request, response) => {
+router.patch('/:questionId', checkAuth, async (request, response) => {
   try {
     const updatedQuestion = await QuestionsService.update(
       request.params.questionId,
@@ -44,11 +57,19 @@ router.patch('/:questionId', async (request, response) => {
   }
 });
 
-router.delete('/:questionId', async (request, response) => {
-  const isDeleted = await QuestionsService.delete(request.params.questionId);
-  isDeleted
-    ? response.end()
-    : notFound(request, response)
+router.delete('/:questionId', checkAuth, async (request, response) => {
+
+  try {
+    if (UserRole.Admin == request.headers.role) {
+      const isDeleted = await QuestionsService.delete(request.params.questionId);
+      isDeleted ? response.end() : notFound(request, response)
+    } else {
+      badRequest(request, response, 'This request is not allowed to users with role: USER')
+    }   
+  } catch (error) {
+    badRequest(request, response, error)
+  }
+
 });
 
 module.exports = router;
